requirements.txt
# Core Dependencies
playwright>=1.40.0
pytest>=7.4.0
pyyaml>=6.0
openpyxl>=3.1.0
apscheduler>=3.10.0

# Optional - for enhanced reporting
allure-pytest>=2.13.0
jinja2>=3.1.0

# For API testing
requests>=2.31.0
httpx>=0.25.0

# Utilities
python-dateutil>=2.8.0
